+++ Tests/Unit/AnalyticsManagerSpec.swift
@@ -667,3 +667,6 @@ final class AnalyticsManagerSpec: QuickSpec {
         }
     }
 }
+
+// swiftlint:enable type_body_length
+// swiftlint:enable function_body_length