+++ Tests/Unit/RAnalyticsDatabaseSpec.swift
@@ -511,3 +511,7 @@ class RAnalyticsDatabaseSpec: QuickSpec {
         }
     }
 }
+
+// swiftlint:enable type_body_length
+// swiftlint:enable function_body_length
+// swiftlint:enable line_length