+++ Tests/Unit/RAnalyticsExternalCollectorSpec.swift
@@ -579,3 +579,6 @@ final class RAnalyticsExternalCollectorSpec: QuickSpec {
         }
     }
 }
+
+// swiftlint:enable type_body_length
+// swiftlint:enable function_body_length