+++ Tests/Unit/RAnalyticsRATTrackerPayloadSpec.swift
@@ -1323,3 +1323,8 @@ class RAnalyticsRATTrackerPayloadSpec: QuickSpec {
         }
     }
 }
+
+// swiftlint:enable line_length
+// swiftlint:enable type_body_length
+// swiftlint:enable function_body_length
+// swiftlint:enable control_statement