+++ Tests/Unit/ReferralAppModelSpec.swift
@@ -558,3 +558,7 @@ final class ReferralAppModelSpec: QuickSpec {
         }
     }
 }
+
+// swiftlint:enable line_length
+// swiftlint:enable type_body_length
+// swiftlint:enable function_body_length