+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * [CONRAT-28527](https://jira.rakuten-it.com/jira/browse/CONRAT-28527): Replace ActionParameters by GeoActionParameters.
 * [CONRAT-28251](https://jira.rakuten-it.com/jira/browse/CONRAT-28251): Move GeoTracker from GeoManager to AnalyticsManager.
 * [CONRAT-28597](https://jira.rakuten-it.com/jira/browse/CONRAT-28597): Refactor GeoConfiguration.
+* [CONRAT-28751](https://jira.rakuten-it.com/jira/browse/CONRAT-28751): Process loc event with AnalyticsManager in order to add additional state parameters.
 
 ### Changes
 * [CONRAT-27817](https://jira.rakuten-it.com/jira/browse/CONRAT-27817): Remove dependency on RSDKUtils.